♻️ refactor: change file directory & change name (#151)
파일 디렉토리 및 파일 이름 변경

#151
